+++ src/main/java/com/fnrsportfolio/negraeportfolio/repository/EducacionRepository.java
@@ -6,10 +6,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EducacionRepository extends JpaRepository<Educacion, Integer> {
+public interface EducacionRepository extends JpaRepository<Educacion, Long> {
 /*
     public Optional<Educacion> findByNombreEdu(String nombreEdu);
 
     public boolean existByNombreEdu(String nombreEdu);
-    */
+*/
 }
